audio: Add smpa backend to SND_DEVICE_OUT_SPEAKER_EXTERNAL_[1,2].
It turns out, that in some cases, the device can't convinced to
play sound out of the speaker. A deeper look revealed some differences
in log output, when preparing the audio device.

Playing correctly:
08-10 21:30:09.716  4384  4421 I audio_hw_utils: send_app_type_cfg_for_device PLAYBACK app_type 69936, acdb_dev_id 14, sample_rate 48000, snd_device_be_idx 43
No sound:
08-10 21:27:31.499   581  9931 I audio_hw_utils: send_app_type_cfg_for_device PLAYBACK app_type 69936, acdb_dev_id 14, sample_rate 48000, snd_device_be_idx 2

After adding some debugging it turns out, the following
...
08-11 09:31:03.497   572  2509 E audio_hw_utils: send_app_type_cfg_for_device: platform_get_snd_device_name(2)=speaker, acbdid=14, be_idx=43
08-11 09:31:03.501   572  2509 E audio_hw_utils: send_app_type_cfg_for_device: platform_get_snd_device_name(3)=speaker-ext-1, acbdid=14, be_idx=2
08-11 09:31:03.503   572  2509 E audio_hw_utils: send_app_type_cfg_for_device: platform_get_snd_device_name(4)=speaker-ext-2, acbdid=14, be_idx=2
...

idx=43 is set to speaker (attributed with backend="smpa" in audio_platform_info.xml),
where idx=2 (and the acbdid=14) leads to speaker-ext-1 and speaker-ext-2 which doesn't
provide the backend="smpa" attribute in their definition.

As far as I know, the missing backend="smpa" attribute leads the audio hal
to choose the wrong mixer path, which is in the correctly working case:
08-10 21:30:09.717  4384  4421 D audio_hw_primary: enable_audio_route: apply mixer and update path: deep-buffer-playback smpa

but  in the no sound case:
08-10 21:27:31.509   581  9931 D audio_hw_primary: enable_audio_route: apply mixer and update path: deep-buffer-playback

Adding the backend attribute to the missing ones, should solve the problem for these use cases now.

(BTW: It's currently unclear, why the audio hal is choosing speaker-ext-1/=speaker-ext-2 instead of speaker.)

Change-Id: Ib6fdf40807b6bdf07099ba7f451fb12ea8f17c9d